feat: my vote history table data source update
