+++ build.config/common.js
@@ -18,9 +18,6 @@ module.exports = {
       },
     ],
   },
-  typescript: {
-    ignoreBuildErrors: true,
-  },
   productionBrowserSourceMaps: true,
   webpack: (config, { webpack }) => {
     config.module.rules.push({
