+++ src/app/network-dao/proposal-list/page.tsx
@@ -1,14 +1,10 @@
 'use client';
+import React from 'react';
+// import dynamic from 'next/dynamic';
+// export default dynamic(() => import('./_page'), { ssr: false });
 
-import { useParams } from 'next/navigation';
-import DaoDetail from '../../dao/[daoId]/_page'
-import qs from 'query-string';
-interface IProps {
-}
-export default function DeoDetails(props: IProps) {
-  const { daoId } = qs.parse(window.location.search);
-  return (
-    <div>
-    <DaoDetail daoId={daoId as string} />
-  </div>)
+export default function NetworkDaoHome() { 
+    return (
+        <div>123</div>
+    )
 }
